cryptohash/src/commonTest/kotlin/com/appmattus/crypto/internal/Haraka512_256Test.kt
/*
 * Copyright 2021 Appmattus Limited
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

@file:Suppress("ClassName")

package com.appmattus.crypto.internal

import com.appmattus.crypto.Algorithm
import com.appmattus.crypto.Digest
import com.appmattus.crypto.internal.core.sphlib.testKatHex
import kotlin.test.AfterTest
import kotlin.test.BeforeTest
import kotlin.test.Test
import kotlin.test.assertNotNull
import kotlin.test.assertNull

class Haraka512_256CoreTest : Haraka512_256Test() {
    override fun digest(): Digest<*> = CoreDigest.create(Algorithm.Haraka512_256)

    @Test
    fun hasImplementation() {
        assertNotNull(digest())
    }
}

// No built-in support
class Haraka512_256InstalledProviderTest {

    @BeforeTest
    fun beforeTest() {
        installPlatformProvider()
    }

    @AfterTest
    fun afterTest() {
        removePlatformProvider()
    }

    @Test
    fun noImplementation() {
        assertNull(PlatformDigest().create(Algorithm.Haraka512_256))
    }
}

/**
 * Test Haraka-512-256 (aka Haraka-512 v2) implementation.
 */
abstract class Haraka512_256Test {

    abstract fun digest(): Digest<*>

    // From https://eprint.iacr.org/2016/098.pdf
    @Test
    fun testHaraka512_256() {
        testKatHex(
            digest(),
            "000102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f202122232425262728292a2b2c2d2e2f303132333435363738393a3b3c3d3e3f",
            "be7f723b4e80a99813b292287f306f625a6d57331cae5f34dd9277b0945be2aa"
        )
    }
}
